pyproject.toml
[project]
name = "llmops-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "injector>=0.24.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.7",
    "openai>=2.15.0",
    "psycopg2>=2.9.11",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.11",
]
